+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ## Added
 
+- Add `has_consent_terms` boolean field to `Order` model
 - Bind terms and conditions to the contract definition template
 - Extend Site model to store terms and conditions
 - Add `join_and` and `list_key` template tags